// client/src/types/enums.tsx
export enum COGNITIVE_FUNCTIONS {
    FI = "Fi",
    FE = "Fe",
    TI = "Ti",
    TE = "Te",
    NI = "Ni",
    NE = "Ne",
    SI = "Si",
    SE = "Se"
}

export enum MBTI_TYPES {
    INTJ = "intj", INTP = "intp", ENTJ = "entj", ENTP = "entp",
    INFJ = "infj", INFP = "infp", ENFJ = "enfj", ENFP = "enfp",
    ISTJ = "istj", ISFJ = "isfj", ESTJ = "estj", ESFJ = "esfj",
    ISTP = "istp", ISFP = "istp", ESTP = "estp", ESFP = "esfp"
}

export enum THEME_TYPES {
    LIGHT = "theme-light",
    DARK = "theme-dark",
    WARM = "theme-warm"
}